fix: upgrade SpacetimeDB CLI to 2.0.5 and regenerate bindings
- Regenerate module_bindings with CLI 2.0.5 (adds required accessor field to indexes)
- Fix stdb blog page object: use select click instead of input fill for category field

+++ backend/spacetimedb/module_bindings/invalidate_movie_reducer.ts
@@ -9,5 +9,5 @@ import {
   type Infer as __Infer,
 } from "spacetimedb";
 export default {
-  tmdbId: __t.f64(),
+  tmdbId: __t.u32(),
 };
